Fixed issue where JMP_REL and JMP_REL_SHORT and JCC_REL operand was stored as unsigned this casued issues if a trampoline had a negative offset.

+++ src/hook.c
@@ -410,13 +410,13 @@ static MH_STATUS EnableHookLL(UINT pos, BOOL enable)
     {
         PJMP_REL pJmp = (PJMP_REL)pPatchTarget;
         pJmp->opcode = 0xE9;
-        pJmp->operand = (UINT32)((LPBYTE)pHook->pDetour - (pPatchTarget + sizeof(JMP_REL)));
+        pJmp->operand = (LPBYTE)pHook->pDetour - (pPatchTarget + sizeof(JMP_REL));
 
         if (pHook->patchAbove)
         {
             PJMP_REL_SHORT pShortJmp = (PJMP_REL_SHORT)pHook->pTarget;
             pShortJmp->opcode = 0xEB;
-            pShortJmp->operand = (UINT8)(0 - (sizeof(JMP_REL_SHORT) + sizeof(JMP_REL)));
+            pShortJmp->operand = 0 - (sizeof(JMP_REL_SHORT) + sizeof(JMP_REL));
         }
     }
     else
